pyproject.toml
[tool.poetry]
name = "panda_tg_bot"
version = "0.1.0"
description = "my silly telegram bot"
authors = ["Pandaaaa906 <ye.pandaaaa906@gmail.com>"]
license = "GPL-3.0+"

[tool.poetry.dependencies]
python = "^3.6"
Telethon = "^1.7"
"hurry.filesize" = "^0.9.0"
Pillow = "^6.0"
cryptg = "^0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
my-script = 'tg_sticker_bot:run'